+++ services/ipc/src/user_auth_service.cpp
@@ -132,7 +132,7 @@ void UserAuthService::GetProperty(std::optional<int32_t> userId, AuthType authTy
         return;
     }
     Attributes attr;
-    attr.SetUint32Value(Attributes::ATTR_AUTH_TYPE, static_cast<uint32_t>(authType));
+    attr.SetInt32Value(Attributes::ATTR_AUTH_TYPE, authType);
     attr.SetUint32Value(Attributes::ATTR_PROPERTY_MODE, PROPERTY_MODE_GET);
     attr.SetUint64Value(Attributes::ATTR_TEMPLATE_ID, templateId);
     attr.SetUint64Value(Attributes::ATTR_CALLER_UID, static_cast<uint64_t>(this->GetCallingUid()));
